PM2 metrics moved to "pm2-metric.js"
Now we have 2 files, 1 is entire of http endpoints , the second is just
gives data;
